+++ .graphqlrc.ts
@@ -1,0 +1,4 @@
+export default {
+  schema: 'libs/graphql-schema/generated/complete-schema.graphql',
+  documents: 'libs/graphql-schema/src/apollo/*.gql',
+}