+++ libs/graphql-schema/src/index.ts
@@ -1,9 +1,7 @@
-import { DIRECTIVES as directives } from '@graphql-codegen/typescript-mongodb'
-
 import { EffectReasonType } from './effect-reason-type'
 import { MoveType } from './move-type'
 import RequestedFields from './requested-fields'
 import scalars from './scalars'
-import schema from './schema'
+import { typeDefs } from '../generated/typeDefs'
 
-export { directives, EffectReasonType, MoveType, scalars, schema, RequestedFields }
+export { EffectReasonType, MoveType, RequestedFields, typeDefs, scalars }