+++ msvs/ui/libs/client/src/components/DeckEditor.tsx
@@ -1,26 +1,28 @@
 import { Dispatch, PropsWithChildren, SetStateAction, useState } from 'react'
+import { useMutation, useQuery } from '@apollo/client'
 
+import {
+  AddDeckDocument,
+  Combat,
+  Deck,
+  DecksQuery,
+  DecksDocument,
+  DeckUnit,
+  DlcKey,
+  EffectKey,
+  Faction,
+  FactionKey,
+  FactionsDocument,
+  FactionsQuery,
+  LeadersDocument,
+  Unit,
+  UnitsDocument,
+  UnitStats,
+} from '@gwent/graphql-schema/apollo-typings'
 import addToCacheList from '../util/add-to-cache-list'
 import Centered from '../components/Centered'
 import CloseButton from './CloseButton'
 import { combineUnitStats, sortObjectArray } from '@gwent/utils'
-import { DeckUnit, EffectKey } from '@gwent/graphql-schema/resolver-typings'
-import {
-  DecksQuery,
-  DecksDocument,
-  FactionKey,
-  useAddDeckMutation,
-  useFactionsQuery,
-  useLeadersQuery,
-  useUnitsQuery,
-  Unit,
-  Combat,
-  UnitStats,
-  DlcKey,
-  Faction,
-  Deck,
-  FactionsQuery,
-} from '@gwent/graphql-schema/apollo-typings'
 import DlcTag from '../components/DlcTag'
 import {
   FILTERS,
@@ -57,7 +59,7 @@ export default function DeckEditor({ deck, onCancel, onSave }: DeckEditorProps) 
   const [selectedUnits, setSelectedUnits] = useState<DeckUnit[]>([])
   const [deckUnits, setDeckUnits] = useState<DeckUnit[]>([])
   const { checkAuth } = useUserContext()
-  const [addDeck, { loading, error }] = useAddDeckMutation({
+  const [addDeck, { loading, error }] = useMutation(AddDeckDocument, {
     variables: {
       faction: faction?.key as FactionKey,
       leader: leaderId as string,
@@ -209,7 +211,7 @@ function renderNameAndFaction({
     error: factionsError,
     data: factionsData,
     refetch: factionsRefetch,
-  } = useFactionsQuery({
+  } = useQuery(FactionsDocument, {
     onError: (error) => {
       checkAuth(error, factionsRefetch)
     },
@@ -426,7 +428,7 @@ function renderLeader({
     error: leadersError,
     data: leadersData,
     refetch: leadersRefetch,
-  } = useLeadersQuery({
+  } = useQuery(LeadersDocument, {
     variables: {
       factions: faction ? [faction.key] : [],
     },
@@ -591,7 +593,7 @@ function renderUnits({
     loading: factionUnitsLoading,
     error: factionUnitsError,
     refetch: factionUnitsRefetch,
-  } = useUnitsQuery({
+  } = useQuery(UnitsDocument, {
     variables: {
       deckable: true,
       factions: faction ? [faction.key] : [],
@@ -616,7 +618,7 @@ function renderUnits({
     loading: neutralUnitsLoading,
     error: neutralUnitsError,
     refetch: neutralUnitsRefetch,
-  } = useUnitsQuery({
+  } = useQuery(UnitsDocument, {
     variables: {
       deckable: true,
       factions: [FactionKey.Neutral],