+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.15.1 - 2025-08-29
+
+### Fixed
+
+- Replace @graphql-codegen/typescript-react-apollo with @graphql-codegen/client-preset ([#43](https://github.com/aboe026/gwent/pull/43)) (Adam Boe)
+- Syntax support for GraphQL ([#43](https://github.com/aboe026/gwent/pull/43)) (Adam Boe)
+
 ## 0.15.0 - 2025-08-27
 
 ### Added